+++ iivs/common/data/__init__.py
@@ -5,6 +5,7 @@ __all__ = (
     "Timestamp",
     "TimestampSequence",
     "TimestampsFixedFPS",
+    "ValueRangeMixin",
     "read_npy_shape",
     "validate_float32_array",
     "validate_float_array",
@@ -13,7 +14,7 @@ __all__ = (
     "write_npy",
 )
 
-from iivs.common.data.mixin import FrameShapedMixin
+from iivs.common.data.mixin import FrameShapedMixin, ValueRangeMixin
 from iivs.common.data.npy import read_npy_shape, write_npy
 from iivs.common.data.sequence import ArrayFileList
 from iivs.common.data.timestamp import Timestamp, TimestampSequence, TimestampsFixedFPS